Reduce mobile layout shift
Previously, in mobile screens, the full catalog and article trays
would display prior to the data being loaded.  After the data is
loaded, the catalog tray is almost always so big that it pushes
the article tray below the fold in mobile views.  This layout
shift could be annoying to users with slow connections who see
articles momentarily, then it disappears.

Setting a reasonable min-height for each tray makes this layout
shift less drastic, and brings the mobile CLS (as measured locally
with Lighthouse using the search term "music") from .36 to .143.