Include aria-label for sections
Connected to #370

This successfully adds an aria-label for all sections, but they do not seem to display in the rotor, but it's possible I'm doing this wrong, since the directions in the ticket to replicate do not work for me. I may be getting to the list of links, rather than the landmarks menu?